refactor(directory structure): add `agora.js` and `utils.js` in `src/modules/`
move functions from `index.js` to files above

BREAKING CHANGE: **directory structure has changed!**
